native: add CMakeLists.txt
Invoke cargo through add_custom_command() and provide a target
the flutter project can link to.

Android is using an older version of cmake as of now
which does not support using generator expressions for OUTPUT,
so we have to check for CMAKE_BUILD_TYPE.

The provided target, mpc_sigs, is an interface library.
Using an imported library would seem to make more sense,
but this setup failed to create a relocatable bundle
(new to cmake :(