tasks: remove long-running workers
Send protocol context to the main isolate after each update.

The disadvantages are that more data needs to be sent between
the main and worker isolate, isolate startup may take additional
time.
On the flipside, this brings easier state management and enables
persistence in the future (context is not recoverable from the
server and must therefore be saved in each round).