file_store: use device-specific directories
For testing purposes, it will be advantageous to be able to
control multiple devices/users from a single process, so to
avoid conflicts create a separate dir for each of them.

(as the rest of the app is currently designed for one user only,
use Uuid([]) for the device id to avoid more code changes).