split MpcModel, inject repositories
Inject repositories using AppContainer.
This is currently rather messy as the repositories depend on the
currently signed in user/current server. If multi-user support is
desired, one could create e.g. a Session holding the current user
and corresponding repos.

Sync class takes over the responsibility of periodically syncing
the repositories with the server.